Move document initialisation to document plugin
