+++ Resources/Private/Templates/Document/Main.html
@@ -3,6 +3,6 @@
       data-namespace-typo3-fluid="true"
       lang="en">
 
-Document
+    <script>{docConfiguration -> f:format.raw()}</script>
 
 </html>
